+++ src/mulle-bootstrap-mv-force.sh
@@ -1,0 +1,69 @@
+#! /usr/bin/env bash
+#
+#   Copyright (c) 2017 Nat! - Mulle kybernetiK
+#   All rights reserved.
+#
+#   Redistribution and use in source and binary forms, with or without
+#   modification, are permitted provided that the following conditions are met:
+#
+#   Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+#   Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+#
+#   Neither the name of Mulle kybernetiK nor the names of its contributors
+#   may be used to endorse or promote products derived from this software
+#   without specific prior written permission.
+#
+#   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+#   AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+#   IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+#   ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+#   LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+#   CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+#   SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+#   INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+#   CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+#   ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+#   POSSIBILITY OF SUCH DAMAGE.
+#
+
+#
+# remove destination directory if it exists
+#
+main()
+{
+   local arguments
+
+   while [ $# -gt 2 ]
+   do
+      arguments="`concat "${arguments}" "'$1'"`"
+      shift
+   done
+
+   local src
+   local dst
+
+   src="$1"
+   dst="$2"
+
+   local name
+
+   name="`basename -- "${src}"`"
+   if [ -d "${dst}/${name}" ]
+   then
+      exekutor rm -rf "${dst}/${name}"
+   fi
+
+   eval_exekutor "mv -f ${arguments} '${src}' '${dst}'"
+
+}
+
+LIBEXEC_PATH="`dirname -- "$0"`"
+
+. "${LIBEXEC_PATH}/mulle-bootstrap-logging.sh"
+. "${LIBEXEC_PATH}/mulle-bootstrap-functions.sh"
+
+main "$@"