+++ src/mulle-bootstrap-scripts.sh
@@ -51,7 +51,7 @@ run_script()
          env | sed 's/^\(.\)/   \1/' >&2
          echo "}" >&2
       fi
-      exekutor "${script}" "$@" >&2 || fail "script \"${script}\" did not run successfully"
+      exekutor "${script}" "$@" >&2 || return 1
    else
       if [ ! -e "${script}" ]
       then